# Optimize point in ring